+++ tests/unit/test_config_utils.py
@@ -11,62 +11,12 @@ configuration files, the expected behavior of configuration updates, and the
 correctness of the underlying logic in the config_utils module.
 """
 
-import shutil
 
 import pytest
 
 from cytosnake.utils.config_utils import load_configs, update_config
 
 
-# ---------------
-# PyTest Fixtures
-# ---------------
-@pytest.fixture
-def config_file_test(tmp_path: pytest.fixture, request: pytest.fixture):
-    """
-    Generates a path to config file
-
-    Parameters
-    ----------
-    tmp_path : pytest.fixture
-        temporary path generated by pytest fixures
-
-    request : pytest.fixture
-        provides access to information and resources related to the currently
-        executing test
-
-    Return
-    ------
-    pathlib.Path
-        Path to testing directory.
-
-    Note
-    ----
-    Paths generated are teared down after the test is finished.
-    """
-
-    # creating temporary directory
-    unit_test_name = request.node.name
-    test_dir = tmp_path / f"unit-testing-{unit_test_name}"
-    test_dir.mkdir()
-
-    # copy any config files into the testing directory
-    shutil.copy("../datasets/test_config.yaml", test_dir)
-
-    # tear down of the directory after the test is done
-    def cleanup():
-        """removes testing directory after finishing"""
-        shutil.rmtree(test_dir)
-
-    request.addfinalizer(cleanup)
-
-    # return path to testing directory
-    return test_dir / "test_config.yaml"
-
-
-# --------------
-# Configs tests
-# --------------
 @pytest.mark.positive
 def test_add_new_entry(config_file_test):
     """Postive test for entering a new config entry into a config file"""
